include/l4/can.h

+++ src/kern/arch/x86/include/x86/io.h
@@ -2,38 +2,38 @@
 
 #include <inttypes.h>
 
-static uchar inb(ushort port)
+static uint8_t inb(uint16_t port)
 {
-	uchar data;
+	uint8_t data;
 	asm volatile ("in %%dx, %%al\n" : "=a" (data) : "d" (port));
 	return data;
 }
 
-static void outb(ushort port, uchar data)
+static void outb(uint16_t port, uint8_t data)
 {
 	asm volatile ("out %%al, %%dx\n" :: "a" (data), "d" (port));
 }
 
-static ushort inw(ushort port)
+static uint16_t inw(uint16_t port)
 {
-	ushort data;
+	uint16_t data;
 	asm volatile ("in %%dx, %%ax\n" : "=a" (data) : "d" (port));
 	return data;
 }
 
-static void outw(ushort port, ushort data)
+static void outw(uint16_t port, uint16_t data)
 {
 	asm volatile ("out %%ax, %%dx\n" :: "a" (data), "d" (port));
 }
 
-static ulong inl(ushort port)
+static uint32_t inl(uint16_t port)
 {
-	ulong data;
+	uint32_t data;
 	asm volatile ("in %%dx, %%eax\n" : "=a" (data) : "d" (port));
 	return data;
 }
 
-static void outl(ushort port, ulong data)
+static void outl(uint16_t port, uint32_t data)
 {
 	asm volatile ("out %%eax, %%dx\n" :: "a" (data), "d" (port));
 }
